Sed-less pure ruby obfuscation solution, refactoring secrets
