+++ gcp/rakefiles/vars.rb
@@ -1,9 +1,9 @@
-require "securerandom"
 require "yaml"
 
 class Vars
 
-  SENSITIVE = "<sensitive>"
+  # Hack to avoid changes in gpii-version-updater
+  VERSION_FILE = '../../../aws/modules/deploy/version.yml'
 
   def self.set_vars(env, project_type)
     if ["dev"].include?(env)
@@ -53,8 +53,7 @@ class Vars
     end
 
     # Hack to avoid changes in gpii-version-updater
-    version_file = '../../../aws/modules/deploy/version.yml'
-    versions = YAML.load(File.read(version_file))
+    versions = YAML.load(File.read(Vars::VERSION_FILE))
     if versions['flowmanager']
       ENV['TF_VAR_flowmanager_repository'] = versions['flowmanager'].split('@')[0]
       ENV['TF_VAR_flowmanager_checksum'] = versions['flowmanager'].split('@')[1]
@@ -66,44 +65,6 @@ class Vars
     if versions['gpii-dataloader']
       ENV['TF_VAR_dataloader_repository'] = versions['gpii-dataloader'].split('@')[0]
       ENV['TF_VAR_dataloader_checksum'] = versions['gpii-dataloader'].split('@')[1]
-    end
-  end
-
-  def self.set_secrets()
-    saved_secrets_file_path = "../#{ENV['ENV']}/secrets/#{ENV["TF_VAR_project_id"]}-secrets.yml"
-    sed_cfg_file_path = "../#{ENV['ENV']}/secrets/#{ENV["TF_VAR_project_id"]}-sed.cfg"
-    sed_cfg = ""
-
-    begin
-      @secrets = YAML.load(File.read(saved_secrets_file_path))
-    rescue Errno::ENOENT
-      generate_file = true
-      @secrets = Hash.new
-    end
-
-    [ \
-      'couchdb_admin_username', \
-      'couchdb_admin_password', \
-      'couchdb_secret', \
-    ].each do |secret|
-      unless ENV[secret.upcase].to_s.empty?
-        @secrets[secret] = ENV[secret.upcase]
-        # we don't want to store Environment variables
-        generate_file = false
-      end
-      @secrets[secret] = SecureRandom.hex if @secrets[secret].to_s.empty?
-      ENV["TF_VAR_#{secret}"] = @secrets[secret]
-      sed_cfg << "s/#{@secrets[secret]}/#{Vars::SENSITIVE}/g\n"
-    end
-
-    if generate_file
-      puts "Secret file #{saved_secrets_file_path} for this deployment not found. I will create one."
-      File.open(saved_secrets_file_path, 'w+') do |file|
-        file.write(@secrets.to_yaml)
-      end
-      File.open(sed_cfg_file_path, 'w+') do |file|
-        file.write(sed_cfg)
-      end
     end
   end
 end
